+++ crates/cli/src/cmd/mod.rs
@@ -1,0 +1,5 @@
+pub mod build;
+pub mod check;
+pub mod new;
+pub mod tx;
+pub mod uplc;
