--- crates/cli/src/lib.rs
@@ -1,2 +1,0 @@
-pub use aiken_lang;
-pub use uplc;
